test case fix for dead lettter queue
